Finally figured out the problem - k,v as references
